rename files without jsx to .ts extension
