+++ src/redux/exclusiveItems/ExclusiveItems.ts
