+++ src/redux/exclusiveItems/defaultState.ts
